+++ formguard/__init__.py
@@ -1,3 +1,3 @@
 """Invisible, layered form protection for Django."""
 
-__version__ = '0.2.1'
+__version__ = '0.3.0'